Cleaned up a bit
